requirements.txt
Flask
Flask-JWT-Extended
psycopg2-binary
Pillow
python-dotenv
pytesseract
passlib
requests
Faker